patches: Define strtof_l and strtod_l in locale
When compiled for Aarch64, apps using libcxx fail due to strtof_l and
strtod_l not being defined in include/locale.

This commit adds a patch which includes the header that contains
the definitions in include/locale.

It also modifies Makefile.uk to add the file in the build process.

Signed-off-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>