include/math.h: Add libcxx math.h header
When building `libcxx` with musl, some undefined functions errors appear
because the musl math.h header is used.

`cmath` looks for some funtions defined in an anonymous namespace, but they
are defined as macros within the `musl` headers. This leads to errors since
the macros can not match the required functions.

The functions are defined in `libcxx` headers, but they are not processed
since the `musl` headers are prioritised. We cannot prioritise the `libcxx`
headers since it leads to a lot of issues similar to this one.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>